school_management_system/requirements.txt
Django>=4.2
psycopg2-binary>=2.9
djangorestframework
django-nepali-datetime-field>=0.8.0
django-filter>=25.2
psycopg2>=2.9.11
Faker>=24.0